General Update
First, I changed the time zone to work in Recife/PE/BR. After that, I used makemigrations and migrate to start the server that compiles the information. Then, I went to the admin page and updated it so the table would appear on the site. Finally, I added a search filter, a search bar, and made it visible in a table format.